Remove the YourJobListener class
